Added default cursor.
I was fine with how it was, showing an X over non-interactive windows,
but in practice it shows an X even on programs like mpv and Blender.
While mpv was fine to ignore since I mostly control it with my keyboard,
Blender with an X for a cursor is atrocious. Like, the special cursors
like in sculpt mode do work, but the default does not. I wonder what
that's all about when the normal cursor appears just fine everywhere
else.