Stop using Polybar
I realized I don't actually care to have a top bar with any sort of
information. It's just an eyesore for most of the time that wastes
space. I would rather have one that's a dropdown called with a hotkey
that holds more information than what would fit in a top bar anyway.

An even better approach may be to have a desktop background with even
more information. This would work best for people like me who don't use
desktop icons and just use a black colour for the background.